Add generic option parser for settings refactor V2
This is implemented by giving each setting a type. Based on the type,
the string will be parsed differently.
A generic enum parser is used as well. Instead of having a function for
every enum, only one function is needed. This will take some data with
possible enum values.

All data needed to parse the settings is defined in settings_data.h.
This is useful for keeping all data in one place and it's easier to
check if there are mistakes.